+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## [Unreleased]
 
 - callタグや`fix="true"`を指定したbuttonタグで呼び出されているシナリオファイルに`[return]`タグが存在しない場合、警告として検出するようにしました。[issue#295](https://github.com/orukRed/tyranosyntax/issues/295)
+- `[iscript]`〜`[endscript]`内のJavaScriptで`startTag('jump', ...)`や`startTag('call', ...)`を呼び出しているにもかかわらず、`[endscript]`に`stop="true"`が指定されていない場合、警告として検出するようにしました。[issue#134](https://github.com/orukRed/tyranosyntax/issues/134)
 
 ## [3.0.0] 2026-04-26